Improve check for preferred node
When process was limited to execute on specific CPUs (e.g. via
taskset(1) or numactl(8) cpunodebind/physcpubind option)
check for invalid configuration is limited to nodes assigned to these
CPUs
E.g. for configuration below, where PMEM NUMA nodes are 4,5,6
node   0   1   2   3   4   5   6
  0:  10  21  21  21  17  28  28
  1:  21  10  21  21  28  28  28
  2:  21  21  10  21  28  17  28
  3:  21  21  21  10  28  28  17
  4:  17  28  28  28  10  28  28
  5:  28  28  17  28  28  10  28
  6:  28  28  28  17  28  28  10

Process is eligble to run on CPU which belongs to NUMA nodes: 0,2,3
Process is not eligble to run on NUMA node 1, where three PMEM NUMA
nodes are in same shortest distance to NUMA node 1 - 28